# Optimize string comparisons: Replace list with set membership tests