--- a/sklearn/utils/fixes.py
+++ b/sklearn/utils/fixes.py
@@ -25,7 +25,7 @@ from ..externals._packaging.version import parse as parse_version
 from .parallel import _get_threadpool_controller
 
 _IS_32BIT = 8 * struct.calcsize("P") == 32
-_IS_WASM = platform.machine() in ["wasm32", "wasm64"]
+_IS_WASM = platform.machine() in {"wasm32", "wasm64"}
 
 np_version = parse_version(np.__version__)
 np_base_version = parse_version(np_version.base_version)