--- a/sklearn/utils/multiclass.py
+++ b/sklearn/utils/multiclass.py
@@ -324,7 +324,7 @@ def type_of_target(y, input_name="", raise_unknown=False):
             "Expected array-like (array or non-string sequence), got %r" % y
         )
 
-    sparse_pandas = y.__class__.__name__ in ["SparseSeries", "SparseArray"]
+    sparse_pandas = y.__class__.__name__ in {"SparseSeries", "SparseArray"}
     if sparse_pandas:
         raise ValueError("y cannot be class 'SparseSeries' or 'SparseArray'")